requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiounittest==1.3.0
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asyncio==3.4.3
attrs==23.2.0
azure-cognitiveservices-speech==1.38.0
azure-core==1.30.2
Babel==2.9.1
binaryornot==0.4.4
blinker==1.8.2
botbuilder-core==4.16.1
botbuilder-dialogs==4.16.1
botbuilder-integration-aiohttp==4.16.1
botbuilder-schema==4.16.1
botframework-connector==4.16.1
botframework-streaming==4.16.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.5
cookiecutter==1.7.0
cryptography==43.0.0
datedelta==1.4
distro==1.9.0
emoji==1.7.0
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
future==1.0.0
grapheme==0.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jinja2-time==0.2.0
jsonpickle==1.4.2
MarkupSafe==2.1.5
msal==1.30.0
msrest==0.7.1
multidict==6.0.5
multipledispatch==1.0.0
oauthlib==3.2.2
openai==1.36.1
poyo==0.5.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.4
tqdm==4.66.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==1.26.19
Werkzeug==3.0.3
whichcraft==0.6.1
yarl==1.9.4
